deployTrialApp
